use auth hook for login
