+++ src/pages/Auth/Login.jsx
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useState } from 'react';
-import { getAuthenticatedUser, sendDataToServer } from '../../helpers/Api';
+// import { getAuthenticatedUser, sendDataToServer } from '../../helpers/Api';
 import { Link, useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import FormInput from '../../components/Form/FormInput';
@@ -7,6 +7,7 @@ import Loader from '../../components/ui/Loader';
 import WhatsUppButton from '../../components/Form/WhatsUppButton';
 import HeaderFormLogin from '../../components/Form/HeaderFormLogin';
 import { FaEyeSlash, FaEye } from "react-icons/fa";
+import { useAuth } from '../../hooks/useAuth';
 
 
 
@@ -18,61 +19,85 @@ const Login = () => {
     });
     const [isLoading, setIsLoading] = useState(false);
     const [showPassword, setShowPassword] = useState(false);
+    const { login, checkAuth } = useAuth();
 
     const navigate = useNavigate();
 
 
     useEffect(() => {
-        const checkAuth = async () => {
-            try {
-                const user = await getAuthenticatedUser();
-                if (user) {
-                    if (user.role === 'admin') {
-                        navigate('/admin');
-                    } else {
-                        navigate('/home');
-                    }
-                }
-            } catch (error) {
-                console.log(error);
-            }
-        };
-
         checkAuth();
-    }, [navigate]);
+    }, []);
 
     const handleChange = (event) => {
         const { name, value } = event.target;
-        setFormData(prevState => ({
-            ...prevState,
+        setFormData(prev => ({
+            ...prev,
             [name]: name === "email" ? value.toLowerCase() : value,
         }));
     };
 
-
     const handleLogin = async (e) => {
         e.preventDefault();
         setIsLoading(true);
-
         try {
-            await sendDataToServer('/users/login', formData);
-
-            const user = await getAuthenticatedUser();
-            if (!user) throw new Error('אימות נכשל');
-
-            toast.success('התחברת בהצלחה!');
-
-            if (user.role === 'admin') {
-                navigate('/admin');
-            } else {
-                navigate('/home');
-            }
+            await login(formData);
         } catch (error) {
-            toast.error(error.message || 'שגיאה בהתחברות');
+            console.log(error);
         } finally {
             setIsLoading(false);
         }
     };
+    // useEffect(() => {
+    //     const checkAuth = async () => {
+    //         try {
+    //             const user = await getAuthenticatedUser();
+    //             if (user) {
+    //                 if (user.role === 'admin') {
+    //                     navigate('/admin');
+    //                 } else {
+    //                     navigate('/home');
+    //                 }
+    //             }
+    //         } catch (error) {
+    //             console.log(error);
+    //         }
+    //     };
+
+    //     checkAuth();
+    // }, [navigate]);
+
+    // const handleChange = (event) => {
+    //     const { name, value } = event.target;
+    //     setFormData(prevState => ({
+    //         ...prevState,
+    //         [name]: name === "email" ? value.toLowerCase() : value,
+    //     }));
+    // };
+
+
+    // const handleLogin = async (e) => {
+    //     e.preventDefault();
+    //     setIsLoading(true);
+
+    //     try {
+    //         await sendDataToServer('/users/login', formData);
+
+    //         const user = await getAuthenticatedUser();
+    //         if (!user) throw new Error('אימות נכשל');
+
+    //         toast.success('התחברת בהצלחה!');
+
+    //         if (user.role === 'admin') {
+    //             navigate('/admin');
+    //         } else {
+    //             navigate('/home');
+    //         }
+    //     } catch (error) {
+    //         toast.error(error.message || 'שגיאה בהתחברות');
+    //     } finally {
+    //         setIsLoading(false);
+    //     }
+    // };
 
     const togglePasswordVisibility = useCallback((e) => {
         e.preventDefault();
